camera: introduce strategies
This separates the main camera logic with that of updating the
position and target into a strategy-based approach. This will later
allow defining the TR3 camera as a separate module.
