+++ src/meson.build
@@ -103,6 +103,7 @@ common_sources = [
   'trx/game/anims/commands.c',
   'trx/game/anims/common.c',
   'trx/game/anims/frames.c',
+  'trx/game/camera/box_camera.c',
   'trx/game/camera/cinematic.c',
   'trx/game/camera/common.c',
   'trx/game/camera/fixed.c',
